# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Peres Bakeries Limited API v1",
        "description": "Peres Bakeries Limited API for managing daily transactons in bakery",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/v1/staff/product": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create new product",
                "operationId": "ProductUpload",
                "requestBody": {
                    "description": "Create Product object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProductModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product Successfully created"
                    },
                    "422": {
                        "description": "Unproccessable data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "error": "Unproccessable data"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/v1/staff/product/update/{id}": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Update product details",
                "operationId": "ProductUpdate",
                "requestBody": {
                    "description": "Update product object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProductModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product Record Successfully updated"
                    },
                    "422": {
                        "description": "Unproccessable data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "error": "Unproccessable data"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/v1/staff/product/delete/{id}": {
            "post": {
                "tags": [
                    "Staff"
                ],
                "summary": "Delete a product from storage",
                "operationId": "ProductDelete",
                "responses": {
                    "200": {
                        "description": "Successful delete operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "message": "Product successfully deleted"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/v1/staff/staff": {
            "post": {
                "tags": [
                    "Staff"
                ],
                "summary": "Create new staff",
                "operationId": "StaffUpload",
                "requestBody": {
                    "description": "Create Staff object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateStaffModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Staff Successfully created"
                    },
                    "422": {
                        "description": "Unproccessable data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "error": "Unproccessable data"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/v1/staff/update/{id}": {
            "post": {
                "tags": [
                    "Staff"
                ],
                "summary": "Update staff details",
                "operationId": "StaffUpdate",
                "requestBody": {
                    "description": "Update Staff object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateStaffModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Staff Record Successfully updated"
                    },
                    "422": {
                        "description": "Unproccessable data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "error": "Unproccessable data"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/v1/staff/delete/{id}": {
            "post": {
                "tags": [
                    "Staff"
                ],
                "summary": "Delete a staff from storage",
                "operationId": "StaffDelete",
                "responses": {
                    "200": {
                        "description": "Successful delete operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "message": "Staff successfully deleted"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "CreateProductModel": {
                "title": "CreateProductModel",
                "description": "Pares Bakeries Limited Login Model",
                "properties": {
                    "name": {
                        "title": "Product name",
                        "type": "string",
                        "example": "Sardine Bread"
                    },
                    "price": {
                        "title": "Product price",
                        "type": "string",
                        "example": "200"
                    },
                    "description": {
                        "title": "Product description",
                        "type": "string",
                        "example": "Jumbo size sardine bread"
                    },
                    "shape": {
                        "title": "Product shape",
                        "type": "string",
                        "example": "Cuboid"
                    },
                    "size": {
                        "title": "Product size",
                        "type": "string",
                        "example": "10cm X 10cm X 25cm"
                    },
                    "stock": {
                        "title": "Product quantity in stock",
                        "type": "string",
                        "example": "300"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CreateProductModel"
                }
            },
            "CreateStaffModel": {
                "title": "CreateStaffModel",
                "description": "Pares Bakeries Limited Login Model",
                "properties": {
                    "surname": {
                        "title": "Staff Surname",
                        "type": "string",
                        "example": "Ide"
                    },
                    "firstname": {
                        "title": "Staff First Name",
                        "type": "string",
                        "example": "Peter"
                    },
                    "mobile_number": {
                        "title": "Staff Phone Number",
                        "type": "string",
                        "example": "08012345678"
                    },
                    "address": {
                        "title": "Staff Address",
                        "type": "string",
                        "example": "No. 8 Yabrifa Close, Kpansia, Yenagoa"
                    },
                    "position": {
                        "title": "Staff position",
                        "type": "string",
                        "example": "HR"
                    },
                    "gender": {
                        "title": "Staff Gender",
                        "type": "string",
                        "example": "Male, Female"
                    },
                    "dob": {
                        "title": "Staff Date of Birth",
                        "type": "string",
                        "example": "dd-mm-yyyy, mm-dd-yyyy, yyyy-mm-dd, dd/mm/yyyy, mm/dd/yyyy, yyyy/mm/dd"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CreateStaffModel"
                }
            },
...
        }
    },
...
}